@@ -1,15 +1,16 @@
 name: loader_search_bar
 description: Flutter widget integrating search field feature into app bar.
-version: 1.0.3+3
+version: 2.0.0+200
 author: Tomasz Wyrowinski <tomwyr@hotmail.com>
 homepage: https://github.com/tomwyr/loader-search-bar
 
 dependencies:
   flutter:
     sdk: flutter
+  collection: ^1.15.0-nullsafety.4
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
   flutter_test:
